Refactor quotation controller actions by: - moving some methods from repository to service - creating a tax service pour handle applying tax logic
